chore: move deepEqual to Util
